Some refactor to attempt to fix extra searches
