+++ frontend/src/__generated/index.css
@@ -6656,6 +6656,48 @@ body {
 ._19or1k14:active {
   color: var(--_1pyqka9w);
 }
+.hgp31e0 .ant-modal-content {
+  background-color: var(--_1pyqka92);
+  border-radius: 8px !important;
+  border: var(--_1pyqka9j) solid 1px;
+  box-shadow: 0 6px 12px -2px rgba(59, 59, 59, 0.12);
+}
+.hgp31e0 .ant-modal-body {
+  padding: 0;
+}
+._1gfijho0 {
+  align-items: center;
+  display: flex;
+  height: 20px;
+  justify-content: space-between;
+}
+._1gfijho1 {
+  max-width: 100%;
+}
+._1gfijho2 {
+  color: #6f6e77;
+}
+._1gfijho3 {
+  cursor: pointer;
+  padding-top: 2px;
+  width: 80px;
+}
+._1gfijho4:hover {
+  background: var(--_1pyqka91x);
+}
+._1gfijho5 {
+  background: var(--_1pyqka920);
+  box-shadow: 0 -1px 0 0 rgba(0, 0, 0, 0.1) inset;
+}
+._1gfijho5:hover {
+  background-color: var(--_1pyqka921);
+}
+.tzvxpp0 {
+  height: calc(100vh - var(--header-height));
+}
+.tzvxpp1 {
+  height: calc(100vh - var(--header-height) - var(--banner-height));
+}
 ._1hjkzyj0 {
   padding: 4px;
   gap: 2px;
@@ -6812,15 +6854,6 @@ body {
 ._1m26oyu9:active:enabled {
   background: transparent;
   color: #6f6e77;
-}
-.hgp31e0 .ant-modal-content {
-  background-color: var(--_1pyqka92);
-  border-radius: 8px !important;
-  border: var(--_1pyqka9j) solid 1px;
-  box-shadow: 0 6px 12px -2px rgba(59, 59, 59, 0.12);
-}
-.hgp31e0 .ant-modal-body {
-  padding: 0;
 }
 ._13m092k0 {
   align-items: center;
@@ -7053,39 +7086,6 @@ body {
   height: 20px;
   left: 0;
   transform-origin: left;
-}
-._1gfijho0 {
-  align-items: center;
-  display: flex;
-  height: 20px;
-  justify-content: space-between;
-}
-._1gfijho1 {
-  max-width: 100%;
-}
-._1gfijho2 {
-  color: #6f6e77;
-}
-._1gfijho3 {
-  cursor: pointer;
-  padding-top: 2px;
-  width: 80px;
-}
-._1gfijho4:hover {
-  background: var(--_1pyqka91x);
-}
-._1gfijho5 {
-  background: var(--_1pyqka920);
-  box-shadow: 0 -1px 0 0 rgba(0, 0, 0, 0.1) inset;
-}
-._1gfijho5:hover {
-  background-color: var(--_1pyqka921);
-}
-.tzvxpp0 {
-  height: calc(100vh - var(--header-height));
-}
-.tzvxpp1 {
-  height: calc(100vh - var(--header-height) - var(--banner-height));
 }
 .riaubm0 {
   background: #fdfcfd;
