+++ frontend/src/routers/ProjectRouter/ApplicationRouter.tsx
@@ -5,7 +5,7 @@ import { CanvasPage } from '@pages/Buttons/CanvasV2'
 import ErrorsV2 from '@pages/ErrorsV2/ErrorsV2'
 import IntegrationsPage from '@pages/IntegrationsPage/IntegrationsPage'
 import LogsPage from '@pages/LogsPage/LogsPage'
-import PlayerPage from '@pages/Player/PlayerPage'
+import { PlayerPage } from '@pages/Player/PlayerPage'
 import { SetupRouter } from '@pages/Setup/SetupRouter/SetupRouter'
 import React, { Suspense } from 'react'
 import { Navigate, Route, Routes } from 'react-router-dom'
